LayoutTests/http/tests/security/contentSecurityPolicy/1.1/referrer-origin-http-https-expected.txt

Test that 'origin' works correctly from HTTP to HTTPS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ReferrerTest.referrerResult is "http://127.0.0.1:8000/"
PASS successfullyParsed is true

TEST COMPLETE
